refactor(schemas): migrate Phase 26 endpoint schemas — 2 files, 6 classes + 7 enums (#6042)
- analytics_cfg.py: migrate NodeType, EdgeType, IssueSeverity, IssueType enums
  + AnalyzeRequest→CFGAnalyzeRequest, AnalyzeFileRequest→CFGAnalyzeFileRequest,
  CFGResponse → schemas_analytics.py
- analytics_llm_patterns.py: migrate PromptCategory, OptimizationType, CostLevel
  enums + PromptAnalysisRequest, UsageRecordRequest (with helper method),
  DateRangeParams → schemas_analytics.py

Renames to avoid generic-name collisions:
- AnalyzeRequest → CFGAnalyzeRequest
- AnalyzeFileRequest → CFGAnalyzeFileRequest
